feat(auto-cancel-orders): renamed to cleanup
