--- packages/vendure-plugin-auto-cancel-orders/src/constants.ts
@@ -1,2 +1,0 @@
-export const AUTO_CANCEL_ORDERS_OPTIONS = Symbol('AUTO_CANCEL_ORDERS_OPTIONS');
-export const loggerCtx = 'AutoCancelOrdersPlugin';
