--- packages/vendure-plugin-auto-cancel-orders/src/index.ts
@@ -1,2 +1,0 @@
-export * from './auto-cancel-orders.plugin';
-export * from './service/auto-cancel-orders.service';
