+++ packages/vendure-plugin-order-cleanup/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@pinelab/vendure-plugin-auto-cancel-orders",
+  "name": "@pinelab/vendure-plugin-order-cleanup",
   "version": "1.0.0",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
