+++ packages/vendure-plugin-order-cleanup/src/constants.ts
@@ -1,0 +1,2 @@
+export const ORDER_CLEANUP_OPTIONS = Symbol('ORDER_CLEANUP_OPTIONS');
+export const loggerCtx = 'OrderCleanupPlugin';
