+++ packages/vendure-plugin-order-cleanup/src/index.ts
@@ -1,0 +1,2 @@
+export * from './order-cleanup.plugin';
+export * from './service/order-cleanup.service';
